backup_before_restructure/requirements.txt
# Dependencias del proyecto TCG Web App

# Backend
supabase==2.16.0
python-dotenv==1.1.1
pydantic==2.11.7

# Scraper
requests==2.32.4
beautifulsoup4==4.13.4
lxml==6.0.0
aiohttp==3.9.0
pandas==2.3.0
numpy==2.3.1

# Testing
pytest==8.4.1
pytest-mock==3.14.1

# Utilities
colorama==0.4.6
pygments==2.19.2
